+++ packages/vita-config-eslint/src/index.ts
@@ -1,5 +1,5 @@
 import { baseRule, typescript as tsRules } from "./recommended";
-// import "./setup";
+import "./setup";
 
 /**
  * 参考 UMI 4 方案
@@ -7,7 +7,7 @@ import { baseRule, typescript as tsRules } from "./recommended";
  * 缺点是不支持 type-aware 的 linting 规则（我们通常也不会用他们）
  */
 export default {
-  // parser: require.resolve("@babel/eslint-parser"),
+  parser: require.resolve("@babel/eslint-parser"),
   plugins: ["react", "react-hooks"],
   settings: {
     react: {
@@ -29,25 +29,25 @@ export default {
   //     rules: tsRules,
   //   },
   // ],
-  // parserOptions: {
-  //   ecmaFeatures: {
-  //     jsx: true,
-  //   },
-  //   babelOptions: {
-  //     babelrc: false,
-  //     configFile: false,
-  //     browserslistConfigFile: false,
-  //     presets: [
-  //       [
-  //         require.resolve("@study/vita-preset-babel"),
-  //         {
-  //           env: "development",
-  //           useTypeScript: true,
-  //           enableNewJsxTransform: true,
-  //         },
-  //       ],
-  //     ],
-  //   },
-  //   requireConfigFile: false,
-  // },
+  parserOptions: {
+    ecmaFeatures: {
+      jsx: true,
+    },
+    babelOptions: {
+      babelrc: false,
+      configFile: false,
+      browserslistConfigFile: false,
+      presets: [
+        [
+          require.resolve("@study/vita-preset-babel"),
+          {
+            env: "development",
+            useTypeScript: true,
+            enableNewJsxTransform: true,
+          },
+        ],
+      ],
+    },
+    requireConfigFile: false,
+  },
 };