+++ packages/vita-plugin-build-scripts/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @study/vita-plugin-build-scripts
 
+## 2.0.4-beta.2
+
+### Patch Changes
+
+- 优化 ESLint 配置
+- Updated dependencies
+  - @study/vita-config-eslint@2.0.4-beta.2
+  - @study/vita-preset-babel@2.0.4-beta.2
+
 ## 2.0.4-beta.1
 
 ### Patch Changes